Move windows registry to a separate module
